Moved business logic to the connection manager
